+++ src/biip/_parser.py
@@ -1,15 +1,18 @@
 """The top-level Biip parser."""
 
-from typing import Union
+from typing import Optional, Union
 
 from biip import ParseError
 from biip.gs1 import DEFAULT_SEPARATOR_CHAR, GS1Message
-from biip.gtin import Gtin, GtinFormat
+from biip.gtin import Gtin, GtinFormat, Rcn, RcnRegion
 
 
 def parse(
-    value: str, *, separator_char: str = DEFAULT_SEPARATOR_CHAR
-) -> Union[Gtin, GS1Message]:
+    value: str,
+    *,
+    rcn_region: Optional[RcnRegion] = None,
+    separator_char: str = DEFAULT_SEPARATOR_CHAR,
+) -> Union[Gtin, GS1Message, Rcn]:
     """Identify data format and parse data.
 
     The current strategy is:
@@ -22,6 +25,9 @@ def parse(
 
     Args:
         value: The data to classify and parse.
+        rcn_region: The geographical region whose rules should be used to
+            interpret Restricted Circulation Numbers (RCN).
+            Needed to extract e.g. variable weight/price from GTIN.
         separator_char: Character used in place of the FNC1 symbol.
             Defaults to `<GS>` (ASCII value 29).
             If variable-length fields in the middle of the message are
@@ -37,11 +43,13 @@ def parse(
     try:
         if len(value) in list(GtinFormat):
             try:
-                return Gtin.parse(value)
+                return Gtin.parse(value, rcn_region=rcn_region)
             except ParseError:
                 pass  # Try the next parser
 
-        return GS1Message.parse(value, separator_char=separator_char)
+        return GS1Message.parse(
+            value, rcn_region=rcn_region, separator_char=separator_char
+        )
     except ParseError:
         raise ParseError(
             f"Failed to parse {value!r} as GTIN or GS1 Element String."